# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Productos y Órdenes",
        "description": "Documentación  de la API, incluyendo los endpoints de gestión de productos y órdenes.",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local de desarrollo"
        },
        {
            "url": "https://api.example.com",
            "description": "Servidor de producción"
        }
    ],
    "paths": {
        "/api/ordenes": {
            "get": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Obtener todas las órdenes",
                "description": "Este endpoint devuelve una lista de todas las órdenes registradas. Si no hay órdenes, retorna un mensaje indicando que no se encontraron registros.",
                "operationId": "54e32b626955738721a74e7d5d2ed5c0",
                "responses": {
                    "200": {
                        "description": "Lista de órdenes o mensaje de que no hay órdenes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "description": "Lista de órdenes",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "customer_name": {
                                                        "type": "string",
                                                        "example": "Juan Pérez"
                                                    },
                                                    "customer_email": {
                                                        "type": "string",
                                                        "example": "juan.perez@example.com"
                                                    },
                                                    "total_price": {
                                                        "type": "number",
                                                        "example": 199.99
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "order_date": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-14T12:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "No hay órdenes registradas"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": 200
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Crear una nueva orden",
                "description": "Este endpoint permite crear una nueva orden con información del cliente y el estado inicial.",
                "operationId": "eb9e1a0efee76b2e12c1c3a359280931",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_name": {
                                        "description": "Nombre del cliente",
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "customer_email": {
                                        "description": "Correo electrónico del cliente",
                                        "type": "string",
                                        "example": "juan.perez@example.com"
                                    },
                                    "total_price": {
                                        "description": "Precio total de la orden",
                                        "type": "number",
                                        "example": 199.99
                                    },
                                    "status": {
                                        "description": "Estado inicial de la orden",
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "completed",
                                            "canceled"
                                        ],
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Orden creada con éxito"
                                        },
                                        "order": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "customer_name": {
                                                    "type": "string",
                                                    "example": "Juan Pérez"
                                                },
                                                "customer_email": {
                                                    "type": "string",
                                                    "example": "juan.perez@example.com"
                                                },
                                                "total_price": {
                                                    "type": "number",
                                                    "example": 199.99
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "order_date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-11-14T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error en la validación de datos"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "customer_name": [
                                                    "El campo nombre del cliente es requerido."
                                                ]
                                            }
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}